port: enter and exit hosted game on native 64-bit

+++ src/legacy/client__draw__playerdraw.c
@@ -18,7 +18,7 @@
 extern int nox_win_width;
 extern int nox_win_height;
 
-extern uint32_t dword_8531A0_2576;
+extern uintptr_t dword_8531A0_2576;
 extern uint32_t nox_player_netCode_85319C;
 
 extern uint32_t nox_color_white_2523948;
